backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
pydantic==2.11.7
pydantic-settings==2.10.1
python-multipart==0.0.20
langchain==0.3.27
langgraph==0.6.5
google-genai==1.31.0
openai==1.99.6
pypdf==6.0.0
pytest==8.4.1
httpx==0.28.1
